requirements.txt
Django==3.2.8
djangorestframework==3.12.4
django-filter==21.1
django-environ==0.7.0
drf-yasg==1.20.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-cors-headers==3.9.0
psycopg2==2.9.1
